Initial Project File Upload
This upload includes 2 files. The Raw Data Analysis imports and explores the raw data gathered from SWITRS. The data used in this project was determined and exported in the raw data analysis. The second file Exploratory Analysis and Data Wrangling largely focuses on exploring the data and cleaning the data. Comment and markdown throughout both files include the rationale behind my decisions to handle missing values and manipulate data.