EA and DW + Feature Engineering
This update of EA and DW deals with the missing values for age. Since this was a relatively big chunk of the data, deleting or imputing using the mean, median, or mode would significantly alter the statistical analysis done later in this project. Therefore, the missing age values were imputed using Iterative Imputer. Columns were chosen to predict age and the one hot encoded and scaled. The first time Iterative Imputer was used, it ran all night and did not finish. I reduced the dimensionality of the features using PCA and it drastically helped improve performance. The missing age values were then predicted and rounded to the nearest integer. Lastly, the data types were revised and amended to best fit the appropriate data type. Finally, a cleaned data frame was exported to use in the next notebook.
In the new Feature Engineering file, geopandas was used to get the city and counties for each collision using the coordinates. I used a variety of shapefiles to best capture all cities and counties based on the coordinates. Some coordinates fell outside of a city or past CA borders so sklearn's BallTree was used to choose the nearest city and county using coordinates that had a city and/or county.