EA & DW + FE
In the exploratory analysis and data wrangling file, some columns were added to the drop statements, and the data frame was exported using Parquet to preserve the data types.

In the feature engineering file, city and county information was completely added to each row using scikit learn's BallTree. The imputed data for city and county were then verified using gmplot. An API was obtained to use Google's javascript map API. Although only 5 points appear in the notebook, the data was verified many times. Since this is a paid service, I am limiting the request to only 5 to conserve resources. The data was then split into three data frames for all collisions, at fault, and victims and exported as a CSV and Parquet file for each data frame.